[PL] All json fields now optional

+++ OpenCV/Modules/ResumeList/ViewModel/ResumeListViewModel.swift
@@ -30,6 +30,7 @@ final class ResumeListViewModel: ObservableObject {
 #endif
     @Published var state: LoadingViewState<Resume> = .appeared
     @Published var scannedCode: String = ""
+    @Published var selectedPerson: Person?
 }
 
 extension ResumeListViewModel {
@@ -59,10 +60,12 @@ extension ResumeListViewModel {
     }
     
     // Delete
-    func deleteItem(_ item: Person) {
-        modelContext.delete(item)
-        save()
-        deleteNotes(for: item)
+    func deleteItem(_ item: Person?) {
+        if let item {
+            modelContext.delete(item)
+            save()
+            deleteNotes(for: item)
+        }
     }
     
     private func deleteNotes(for person: Person) {
@@ -79,14 +82,20 @@ extension ResumeListViewModel {
     }
     
     var isValidUrl: Bool {
-        // Check if the URL starts with "https" and ends with ".json"
-        if url.hasPrefix("https://") && url.hasSuffix(".json") {
+        // Check if the URL is valid
+        if isValidURL(url: url) {
             // Check if the URL is a valid URL
-            if let url = URL(string: url), url.scheme == "https" {
+            if let _ = URL(string: url) {
                 return true
             }
         }
         return false
+    }
+    
+    private func isValidURL(url: String) -> Bool {
+        let urlPattern = #"^(https?|ftp)://[^\s/$.?#].[^\s]*$"#
+        let urlTest = NSPredicate(format: "SELF MATCHES %@", urlPattern)
+        return urlTest.evaluate(with: url)
     }
     
     func generateAppClipLink(resumeUrl: String) -> String {
